+++ src/Http/Controllers/DashboardController.php
@@ -50,7 +50,7 @@ class DashboardController extends AdminController implements Listener
      */
     public function missing()
     {
-        throw new NotFoundHttpException("Controller method not found.");
+        throw new NotFoundHttpException('Controller method not found.');
     }
 
     /**
@@ -62,7 +62,7 @@ class DashboardController extends AdminController implements Listener
      */
     public function showDashboard(array $data)
     {
-        set_meta('title', trans("orchestra/foundation::title.home"));
+        set_meta('title', trans('orchestra/foundation::title.home'));
 
         return view('orchestra/foundation::dashboard.index', $data);
     }