+++ src/Http/Controllers/Extension/ViewerController.php
@@ -49,7 +49,7 @@ class ViewerController extends Controller implements Listener
      */
     public function showExtensions(array $data)
     {
-        set_meta('title', trans("orchestra/foundation::title.extensions.list"));
+        set_meta('title', trans('orchestra/foundation::title.extensions.list'));
 
         return view('orchestra/foundation::extensions.index', $data);
     }