+++ docs/changes.md
@@ -5,6 +5,13 @@ title: Foundation Change Log
 
 ## Version 3.0 {#v3-0}
 
+### v3.0.10 {#v3-0-10}
+
+* Add `@set_meta` and `@get_meta` blade helper.
+* Update views to use new helpers.
+* Update `Orchestra\Foundation\Application` to use configurable `databasePath` and `storagePath` properties.
+* Add `Orchestra\Foundation\Http\Middleware\LoginAs`.
+
 ### v3.0.9 {#v3-0-9}
 
 * Add validation rules for `password` and `password_confirmation` for `Orchestra\Foundation\Validation\Account::onRegister()`, useful when validation user creating when password is available.